requirements.txt
pyautogui==0.9.54
PyGetWindow==0.0.9
Pillow==10.4.0
opencv-python==4.10.0.84
